Add the recipe of borsch
At first I found an appropriate image of the borsch's finished version to demonstrate one what they should expect. Then I came up with a short description to let my reader know what this recipe is going to be like. Last but not least, I wrote down the ingredients and steps sections because I wanted to make this process as clear as possible for whoever will decide to give it a try.