Add an alt attribute to the images
I almost forgot to implement this attribute into my project, but it is important! For instance, if my user faces some problems with loading pictures on a webpage, then they will at least be able to read its name. And that short description will help them visualize what is supposed to be there.